backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
rapidfuzz==3.9.0
reportlab==4.2.0
qrcode[pil]==7.4.2
Pillow==10.4.0
google-generativeai==0.8.0
